requirements.txt
flask
flask-socketio
langchain-community
matplotlib
pandas
paho-mqtt
psycopg2
pykafka
sqlalchemy